Refactoring Account, Customer Entity Collection - Functional Requeriments

+++ src/main/java/com/nttdata/account/model/account/SavingsInfo.java
@@ -17,6 +17,6 @@ public class SavingsInfo {
     private BigDecimal interestRate;
 
     @JsonProperty(value = "monthlyMovements")
-    private BigDecimal monthlyMovements;
+    private Integer monthlyMovements;
 
 }
